change file names, move components
